refactor: remove TickerSelectorProvider and consolidate FavoriteButton imports
