+++ src/features/trade/components/FavoriteButton.tsx
